Add intro screen to header
