+++ src/app/page.tsx
@@ -1,6 +1,6 @@
 import { FC } from 'react';
 
-import Header from './components/Header';
+import Header from './components/Header/Header';
 import Nav from './components/Nav';
 import About from './components/About';
 import Tech from './components/Tech';
